Split MuzeiActivity into multiple fragments
Rather than a monolithic MuzeiActivity, split its functionality into three fragments:
- IntroFragment which controls activating Muzei
- TutorialFragment providing a one time tutorial
- ArtDetailFragment as the main view
