+++ main/src/main/res/layout-sw600dp/tutorial_fragment.xml
@@ -16,9 +16,8 @@
 
 <FrameLayout xmlns:android="http://schemas.android.com/apk/res/android"
     android:layout_width="match_parent"
-    android:layout_height="match_parent"
-    android:visibility="gone">
-    <include layout="@layout/muzei_include_tutorial_content"
+    android:layout_height="match_parent">
+    <include layout="@layout/tutorial_content"
         android:layout_width="360dp"
         android:layout_height="300dp"
         android:layout_gravity="center"/>
