+++ main/src/main/res/layout/tutorial_content.xml
